pyproject.toml
[project]
name = "dbttools"
version = "0.1.0"
description = ""
authors = [
    {name = "PaulMullen",email = "paulrpg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core (>=1.10.7,<2.0.0)",
    "dbt-snowflake (>=1.10.0,<2.0.0)",
    "dbt-sqlite (>=1.10.0,<2.0.0)",
    "networkx (>=3.5,<4.0)",
    "plotly (>=6.3.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

